refactor: integrate SeatGridDisplay component for homepage and seat-selection
- Created reusable SeatGridDisplay component with display and interactive modes
- Updated homepage (src/app/page.jsx) to use SeatGridDisplay for seat grid display
- Refactored seat-selection page to use SeatGridDisplay for interactive seat reservation
- Removed deprecated SeatGrid component (no longer needed)
- Eliminates code duplication between homepage and seat-selection pages
- Maintains three-zone layout (left/middle/right) consistent with design requirements
